refactor: extract touch selection into chart module
Isolate two-finger chart selection logic in a dedicated module so chart navigation stays maintainable while preserving touch behavior, and add chart-state tests to increase coverage for updated selection and navigation paths.

Made-with: Cursor
